Add new abstract syntax tree diff checker
This commit adds a new module that checks for differences in the abstract syntax tree (AST) of PHP files. This is done by generating and comparing hashes of the AST for different files or versions of files. The module also provides interfaces and classes for interacting with Git, such as checking if a commit exists, retrieving commit data, or pulling file data from a specific commit. There are also classes for presenting the diffs. The diff information includes the kind of change (addition, modification, deletion) and the status of whether the AST has changed.