Update composer dev dependencies to stable versions
This commit updates the versions of the "phpstan/extension-installer" and "phpunit/phpunit" packages in composer.json to "@stable". This ensures we are using the latest stable versions of these packages that have been thoroughly tested and validated by the community, improving the overall reliability of our application. A "composer.lock" file change comes along with the composer.json file modification, reflecting the updated dependencies.